refactor:move tools to own respective folder
