+++ src/modules/toolbox/components/drawToolContent/index.js
@@ -1,0 +1,6 @@
+
+import { DrawToolContent } from './DrawToolContent';
+
+if (!window.customElements.get(DrawToolContent.tag)) {
+	window.customElements.define(DrawToolContent.tag, DrawToolContent);
+}
