+++ src/modules/toolbox/components/measureToolContent/MeasureToolContent.js
