+++ src/modules/toolbox/components/measureToolContent/index.js
@@ -1,0 +1,5 @@
+import { MeasureToolContent } from './MeasureToolContent';
+
+if (!window.customElements.get(MeasureToolContent.tag)) {
+	window.customElements.define(MeasureToolContent.tag, MeasureToolContent);
+}
