+++ src/modules/toolbox/components/measureToolContent/measureToolContent.css
