+++ src/modules/toolbox/index.js
@@ -1,2 +1,4 @@
 import './components/toolContainer';
+import './components/drawToolContent';
+import './components/measureToolContent';
 import './i18n';
